+++ CHANGELOG.md
@@ -10,10 +10,13 @@
   - Derived enums also automatically generate accurate schemas.
 - Added support for `#[config(serde(...))]` attribute (doesn't support everything, mainly enum
   tagging).
+- Added support for `#[setting(validate)]` on nested fields (was previously an error).
 
 #### 🐞 Fixes
 
 - Fixed a handful of issues with schema generation and partial detection.
+- Fixed an issue where multiline comments weren't parsed correctly.
+- Fixed an issue with string formatting that would incorrectly handle digit characters.
 
 ## 0.10.1